backend/requirements.txt
asgiref==3.5.2
atlassian-python-api==3.23.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
wrapt==1.14.1
